requirements.txt
klayout>=0.28.0
numpy>=1.20.0
pyyaml>=6.0
matplotlib>=3.5.0 